% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readONSData_obsolete.R
\name{discoverONS}
\alias{discoverONS}
\title{Discover datasets provided by ONS through its API}
\usage{
discoverONS(context = NULL)
}
\arguments{
\item{context:}{char,  indicates in which context to start the discovery, either: \code{Census},
\code{Socio-Economic}, \code{Economic}, or \code{Social}. If \code{NULL} it starts showing the 
available contexts.}
}
\description{
Discover datasets provided by ONS through its API
}
