% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readONSData_obsolete.R
\name{readONS}
\alias{readONS}
\title{Retrieve data from ONS}
\usage{
readONS(path = "", context = "", geo = "")
}
\arguments{
\item{path:}{char}

\item{context:}{char,  indicates in which context to start the discovery, either: \code{Census},
\code{Socio-Economic}, \code{Economic}, or \code{Social}. If \code{NULL} it starts showing the 
available contexts.}

\item{geo:}{char}
}
\description{
Retrieve data from ONS
}
