% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth_obsolete.R
\name{setOnsApikey}
\alias{setOnsApikey}
\title{Set your ONS API key in the system}
\usage{
setOnsApikey(apikey)
}
\arguments{
\item{apikey:}{char, API key obtained from ONS}
}
\description{
Set your ONS API key in the system
}
