% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readONSData.R
\name{discoverONS}
\alias{discoverONS}
\title{Discover datasets provided by ONS through its API}
\usage{
discoverONS(start = 0, theApi = getONSapi())
}
\arguments{
\item{start:}{double}
}
\description{
Discover datasets provided by ONS through its API
}
