% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readONSData.R
\name{DiscoverONS}
\alias{DiscoverONS}
\title{Discover datasets provided by ONS through its API}
\usage{
DiscoverONS(start = 0, theApi = GetONSapi())
}
\arguments{
\item{start:}{double}
}
\description{
Discover datasets provided by ONS through its API
}
