% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readONSData.R
\name{discoverONS}
\alias{discoverONS}
\title{Discover datasets provided by ONS through its API}
\usage{
discoverONS(start = 0, theApi = getONSapi())
}
\arguments{
\item{start}{an integer indicating the starting index}

\item{theApi}{a swagger ONS API. Defaults to getONSapi function}
}
\description{
Discover datasets provided by ONS through its API
}
